requirements.txt
absl-py==0.12.0
astroid==2.5.6
astunparse==1.6.3
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
docopt==0.6.2
flatbuffers==1.12
gast==0.3.3
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
isort==5.8.0
joblib==1.0.1
Js2Py==0.71
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.6.0
Markdown==3.3.4
mccabe==0.6.1
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.2.52
opt-einsum==3.3.0
packaging==20.9
Pillow==8.2.0
protobuf==3.17.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyjsparser==2.7.1
pyparsing==2.4.7
PyPrind==2.11.3
pySmartDL==1.3.4
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.4.1
six==1.15.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
toml==0.10.2
tqdm==4.61.0
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.5
Wave==0.0.2
Werkzeug==2.0.1
wget==3.2
wrapt==1.12.1
